# swagger/api-principal.yaml
openapi: 3.0.3
info:
  title: API Principal PhotoSpaces
  description: >-
    Esta API es ofrecida por los microservicios de la aplicación PhotoSpaces, se
    encarga de realizar las acciones, comprobaciones

    y llamadas a las APIs de los servidores de renderizado necesarias para
    satisfacer las peticiones de los clientes.
  version: '1.0'
paths:
  /servers:
    get:
      tags:
        - servers
      summary: Obtener servidores
      description: Obtener servidores registrados en el sistema
      operationId: getServers
      parameters:
        - name: limit
          in: query
          description: Número máximo de servidores que se devolverán por estado
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Todo va bien
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Servers'
        '500':
          description: Se produce un error en la consulta a la base de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetServersServerErrorInfo'
    post:
      tags:
        - servers
      summary: Añadir servidor
      description: Registrar nuevo servidor en el sistema
      operationId: addServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ip:
                  type: string
                  format: ipv4
                  example: 127.0.0.1
                name:
                  type: string
                  example: localhost
      responses:
        '201':
          description: Todo va bien
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '400':
          description: Algún parámetro no válido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddServerUserError'
        '500':
          description: >-
            Error relacionado con la base de datos, con el servidor de
            renderizado, o con la comunicación con este
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddServerServerError'
  '/servers/{id}':
    get:
      tags:
        - servers
      summary: Obtener servidor de renderizado por id
      description: Obtener servidor de renderizado a partir de su oid en la base de datos
      operationId: getServer
      parameters:
        - name: id
          in: path
          description: Id del servidor
          required: true
          schema:
            type: string
            example: 64be9afac8bd803b63efa530
      responses:
        '200':
          description: Todo va bien
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          description: Id no es válido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetServerInvalidIdError'
        '404':
          description: No existe petición con el id especificado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetServerNotFoundError'
        '500':
          description: Error en la consulta a la base de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetServerServerError'
    delete:
      tags:
        - servers
      summary: Eliminar servidor de renderizado
      description: >-
        Desvincular el servidor de renderizado correspondiente al parámetro id
        del sistema
      operationId: deleteServer
      parameters:
        - name: id
          in: path
          description: Id del servidor
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Todo va bien
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '400':
          description: >-
            Id no es válido o se ha enviado la petición se ha enviado en un
            momento ilógico o conflictivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteServerInvalidIdError'
        '404':
          description: No existe servidor con el id especificado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteServerNotFoundError'
        '500':
          description: >-
            Error relacionado con la base de datos, con el servidor de
            renderizado, o con la comunicación con este
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteServerServerError'
  '/servers/{id}/disable':
    post:
      tags:
        - servers
      summary: Deshabilitar servidor de renderizado
      description: Deshabilitar el servidor correspondiente al id especificado
      operationId: disableServer
      parameters:
        - name: id
          in: path
          description: Id del servidor
          required: true
          schema:
            type: string
            example: 64be9afac8bd803b63efa530
      responses:
        '200':
          description: Todo va bien
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisableServerSuccess'
        '400':
          description: >-
            Id no es válido o se ha enviado la petición se ha enviado en un
            momento ilógico o conflictivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisableServerInvalidIdError'
        '404':
          description: No existe servidor con el id especificado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisableServerNotFoundError'
        '500':
          description: >-
            Error relacionado con la base de datos, con el servidor de
            renderizado, o con la comunicación con este
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisableServerServerError'
  '/servers/{id}/enable':
    post:
      tags:
        - servers
      summary: Habilitar servidor de renderizado
      description: Habilitar el servidor correspondiente al id especificado
      operationId: enableServer
      parameters:
        - name: id
          in: path
          description: Id del servidor
          required: true
          schema:
            type: string
            example: 64be9afac8bd803b63efa530
      responses:
        '200':
          description: Todo va bien
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnableServerSuccess'
        '400':
          description: >-
            Id no es válido o se ha enviado la petición se ha enviado en un
            momento ilógico o conflictivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnableServerInvalidIdError'
        '404':
          description: No existe servidor con el id especificado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnableServerNotFoundError'
        '500':
          description: >-
            Error relacionado con la base de datos, con el servidor de
            renderizado, o con la comunicación con este
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnableServerServerError'
  '/servers/{id}/abort':
    post:
      tags:
        - servers
      summary: Abortar procesamiento en servidor de renderizado
      description: >-
        Abortar el procesamiento en el servidor de renderizado correspondiente
        al parámetro id
      operationId: abortServer
      parameters:
        - name: id
          in: path
          description: Id del servidor
          required: true
          schema:
            type: string
            example: 64be9afac8bd803b63efa530
      responses:
        '200':
          description: Todo va bien
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbortServerSuccess'
        '400':
          description: >-
            Id no es válido o se ha enviado la petición se ha enviado en un
            momento ilógico o conflictivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbortServerInvalidIdError'
        '404':
          description: No existe servidor con el id especificado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbortServerNotFoundError'
        '500':
          description: >-
            Error relacionado con la base de datos, con el servidor de
            renderizado, o con la comunicación con este
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbortServerServerError'
...
components:
  schemas:
    GetServersServerErrorInfo:
      type: object
      properties:
        error:
          type: string
          example: Error en la consulta a la base de datos
    Servers:
      type: array
      items:
        type: object
        properties:
          _id:
            type: string
            example: 64be9afac8bd803b63efa530
          name:
            type: string
            example: localhost
          ip:
            type: string
            example: 127.0.0.1
          status:
            type: string
            example: idle
          os:
            type: string
            example: Microsoft Windows 11 Pro
          cpu:
            type: string
            example: Intel® Core™ i7-7700K
          gpu:
            type: string
            example: NVIDIA GeForce GTX 1080 Ti
          blenderVersion:
            type: string
            example: 3.4.1
          registrationDate:
            type: string
            format: date-time
            example: '2023-04-21T15:00:16.398Z'
          timeSpentOnRenderTest:
            type: integer
            format: int64
            example: 56956
          fulfilledRequestsCount:
            type: integer
            format: int64
            example: 0
          enqueuedRequestsCount:
            type: integer
            format: int64
            example: 0
          totalCyclesNeededTime:
            type: integer
            format: int64
            example: 0
          totalCyclesBlenderTime:
            type: integer
            format: int64
            example: 0
          totalCyclesProcessedBytes:
            type: integer
            format: int64
            example: 0
          totalCyclesProcessedPixels:
            type: integer
            format: int64
            example: 0
          cyclesProcessedBytesPerMillisecondOfNeededTime:
            type: integer
            format: int64
            example: 0
          cyclesScore:
            type: integer
            format: int64
            example: null
          totalEeveeNeededTime:
            type: integer
            format: int64
            example: 0
          totalEeveeBlenderTime:
            type: integer
            format: int64
            example: 0
          totalEeveeProcessedBytes:
            type: integer
            format: int64
            example: 0
          totalEeveeProcessedPixels:
            type: integer
            format: int64
            example: 0
          eeveeProcessedBytesPerMillisecondOfNeededTime:
            type: integer
            format: int64
            example: 0
          eeveeScore:
            type: integer
            format: int64
            example: null
          __v:
            type: integer
            format: int64
            example: 0
      example:
        - _id: 64be9afac8bd803b63efa530
          name: localhost
          ip: 127.0.0.1
          status: idle
          os: Microsoft Windows 11 Pro
          cpu: Intel® Core™ i7-7700K
          gpu: NVIDIA GeForce GTX 1080 Ti
          blenderVersion: 3.4.1
          registrationDate: '2023-04-21T15:00:16.398Z'
          timeSpentOnRenderTest: 56956
          fulfilledRequestsCount: 0
          enqueuedRequestsCount: 0
          totalCyclesNeededTime: 0
          totalCyclesBlenderTime: 0
          totalCyclesProcessedBytes: 0
          totalCyclesProcessedPixels: 0
          cyclesProcessedBytesPerMillisecondOfNeededTime: 0
          cyclesScore: null
          totalEeveeNeededTime: 0
          totalEeveeBlenderTime: 0
          totalEeveeProcessedBytes: 0
          totalEeveeProcessedPixels: 0
          eeveeProcessedBytesPerMillisecondOfNeededTime: 0
          eeveeScore: null
          __v: 0
        - _id: 64be9afac8bd803b63efa534
          name: Portátil
          ip: 192.168.1.42
          status: busy
          os: Microsoft Windows 10 Pro
          cpu: Intel® Core™ i7-13700K
          gpu: NVIDIA GeForce RTX 4070
          blenderVersion: 3.0 Alpha
          registrationDate: '2023-04-21T15:00:16.398Z'
          timeSpentOnRenderTest: 10443
          fulfilledRequestsCount: 0
          enqueuedRequestsCount: 0
          totalCyclesNeededTime: 0
          totalCyclesBlenderTime: 0
          totalCyclesProcessedBytes: 0
          totalCyclesProcessedPixels: 0
          cyclesProcessedBytesPerMillisecondOfNeededTime: 0
          cyclesScore: null
          totalEeveeNeededTime: 0
          totalEeveeBlenderTime: 0
          totalEeveeProcessedBytes: 0
          totalEeveeProcessedPixels: 0
          eeveeProcessedBytesPerMillisecondOfNeededTime: 0
          eeveeScore: null
          __v: 0
    Server:
      type: object
      properties:
        _id:
          type: string
          example: 64be9afac8bd803b63efa530
        name:
          type: string
          example: localhost
        ip:
          type: string
          example: 127.0.0.1
        status:
          type: string
          example: idle
        os:
          type: string
          example: Microsoft Windows 11 Pro
        cpu:
          type: string
          example: Intel® Core™ i7-7700K
        gpu:
          type: string
          example: NVIDIA GeForce GTX 1080 Ti
        blenderVersion:
          type: string
          example: 3.4.1
        registrationDate:
          type: string
          format: date-time
          example: '2023-04-21T15:00:16.398Z'
        timeSpentOnRenderTest:
          type: integer
          format: int64
          example: 56956
        fulfilledRequestsCount:
          type: integer
          format: int64
          example: 0
        enqueuedRequestsCount:
          type: integer
          format: int64
          example: 0
        totalCyclesNeededTime:
          type: integer
          format: int64
          example: 0
        totalCyclesBlenderTime:
          type: integer
          format: int64
          example: 0
        totalCyclesProcessedBytes:
          type: integer
          format: int64
          example: 0
        totalCyclesProcessedPixels:
          type: integer
          format: int64
          example: 0
        cyclesProcessedBytesPerMillisecondOfNeededTime:
          type: integer
          format: int64
          example: 0
        cyclesScore:
          type: integer
          format: int64
          example: null
        totalEeveeNeededTime:
          type: integer
          format: int64
          example: 0
        totalEeveeBlenderTime:
          type: integer
          format: int64
          example: 0
        totalEeveeProcessedBytes:
          type: integer
          format: int64
          example: 0
        totalEeveeProcessedPixels:
          type: integer
          format: int64
          example: 0
        eeveeProcessedBytesPerMillisecondOfNeededTime:
          type: integer
          format: int64
          example: 0
        eeveeScore:
          type: integer
          format: int64
          example: null
        __v:
          type: integer
          format: int64
          example: 0
...
    AddServerUserError:
      type: object
      properties:
        error:
          type: string
          example: >-
            No se encontró el nombre del servidor de renderizado en el cuerpo de
            la petición
...
    EnableServerSuccess:
      type: object
      properties:
        message:
          type: string
          example: Servidor habilitado con éxito
...